created all routes

+++ src/app/modules/blog/blog.service.ts
@@ -3,12 +3,13 @@ import { BlogModel } from './blog.model';
 
 import AppError from '../../errors/AppError';
 import httpStatus from 'http-status-codes';
+import QueryBuilder from '../../builder/QueryBuilder';
+import { Types } from 'mongoose';
+import { USER_ROLE } from '../user/user.constant';
 
 const createBlogIntoDB = async (payload: IBlog) => {
-  // Create a new blog entry
   const blog = await BlogModel.create(payload);
 
-  // Populate the 'author' field after creating the blog entry
   const populatedBlog = await BlogModel.findById(blog._id)
     .populate('author', ' -createdAt -updatedAt')
     .select('-isPublished -createdAt -updatedAt');
@@ -17,12 +18,21 @@ const createBlogIntoDB = async (payload: IBlog) => {
 };
 
 const updateBlogIntoDB = async (id: string, payload: Partial<IBlog>) => {
-  const isServiceExists = await BlogModel.findById(id);
-  if (!isServiceExists) {
+  const { author, ...remainingServiceData } = payload;
+  // console.log(payload);
+
+  const isBlogExists = await BlogModel.findById(id);
+  if (!isBlogExists) {
     throw new AppError(httpStatus.NOT_FOUND, 'This blog is not found');
   }
+  // console.log(isBlogExists.author, author);
 
-  const { ...remainingServiceData } = payload;
+  if (isBlogExists.author !== author)
+    throw new AppError(
+      httpStatus.UNAUTHORIZED,
+      'You are not authorized to update this blog',
+    );
+
   const modifiedUpdateData: Record<string, unknown> = {
     ...remainingServiceData,
   };
@@ -37,8 +47,41 @@ const updateBlogIntoDB = async (id: string, payload: Partial<IBlog>) => {
   return result;
 };
 
-const deleteABlogFromDB = async (id: string) => {
+const deleteABlogFromDB = async (
+  id: string,
+  userID: Types.ObjectId,
+  userRole: string,
+) => {
+  const isBlogExists = await BlogModel.findById(id);
+  if (!isBlogExists) {
+    throw new AppError(httpStatus.NOT_FOUND, 'This blog is not found');
+  }
+
+  if (userRole == USER_ROLE.user && isBlogExists.author !== userID)
+    throw new AppError(
+      httpStatus.UNAUTHORIZED,
+      'You are not authorized to delete this blog',
+    );
+
   const result = await BlogModel.deleteOne({ _id: id });
+  return result;
+};
+
+const getAllBlogs = async (query: Record<string, unknown>) => {
+  const queryBuilder = new QueryBuilder(
+    BlogModel.find()
+      .populate('author', '-__v -createdAt -updatedAt')
+      .select('-isPublished -createdAt -updatedAt'),
+    query,
+  );
+
+  const blogQuery = queryBuilder
+    .search(['title', 'content'])
+    .filter()
+    .sort()
+    .paginate();
+
+  const result = await blogQuery.modelQuery;
   return result;
 };
 
@@ -46,4 +89,5 @@ export const BlogService = {
   createBlogIntoDB,
   updateBlogIntoDB,
   deleteABlogFromDB,
+  getAllBlogs,
 };
